service/colplay/requirements.txt
cyclone==1.1
Pillow==3.1.1
python-dateutil==2.5.1
Twisted==16.0.0
webcolors==1.5
